Add internal diskutil package.
